+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 
 - Add `msgpack-value` crate
+- Implement `proptest::arbitrary::Arbitrary` for `msgpack_value::Value` and others
 
 ### Fixed