+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Changed
 
+- Turn `Token` into a non-owning type.
+
 ---
 
 ## 0.5.0 - 2023-03-12